Use pytest fixtures
